requirements.txt
httpx
beautifulsoup4
spacy
pandas
gitpython
openpyxl
gensim
nltk